Move `gen_user_actions` to the codegen module
