+++ lrpar/src/lib/ctbuilder.rs
@@ -37,7 +37,7 @@ use quote::{ToTokens, TokenStreamExt, format_ident, quote};
 use syn::{Generics, parse_quote};
 use wincode::{SchemaRead, SchemaReadOwned, SchemaWrite};
 
-const ACTION_PREFIX: &str = "__gt_";
+pub(crate) const ACTION_PREFIX: &str = "__gt_";
 const GLOBAL_PREFIX: &str = "__GT_";
 const ACTIONS_KIND: &str = "__GtActionsKind";
 const ACTIONS_KIND_PREFIX: &str = "Ak";
@@ -904,7 +904,7 @@ where
         let user_actions = if let YaccKind::Original(YaccOriginalActionKind::UserAction)
         | YaccKind::Grmtools = build_env.yacc_kind()
         {
-            Some(self.gen_user_actions(code_gen, src_env)?)
+            Some(code_gen.gen_user_actions(src_env)?)
         } else {
             None
         };
@@ -1363,145 +1363,6 @@ where
         Ok(wrappers)
     }
 
-    /// Generate the user action functions (if any).
-    fn gen_user_actions(
-        &self,
-        code_gen: &ParserCodegen<LexerTypesT>,
-        src_env: &ParserSrcEnv,
-    ) -> Result<TokenStream, Box<dyn Error>> {
-        let grm = code_gen.grm();
-        let _diag = src_env.yacc_diag();
-        let programs = grm
-            .programs()
-            .as_ref()
-            .map(|s| str::parse::<TokenStream>(s))
-            .transpose()?;
-        let mut action_fns = TokenStream::new();
-        // Convert actions to functions
-        let parsed_parse_generics = make_generics(grm.parse_generics().as_deref())?;
-        let (generics, _, where_clause) = parsed_parse_generics.split_for_impl();
-        let (parse_paramname, parse_paramdef, parse_param_unit);
-        match grm.parse_param() {
-            Some((name, tyname)) => {
-                parse_param_unit = tyname.trim() == "()";
-                parse_paramname = str::parse::<TokenStream>(name)?;
-                let ty = str::parse::<TokenStream>(tyname)?;
-                parse_paramdef = quote!(#parse_paramname: #ty);
-            }
-            None => {
-                parse_param_unit = true;
-                parse_paramname = quote!(());
-                parse_paramdef = quote! {_: ()};
-            }
-        };
-        for pidx in grm.iter_pidxs() {
-            if pidx == grm.start_prod() {
-                continue;
-            }
-
-            // Work out the right type for each argument
-            let mut args = Vec::with_capacity(grm.prod(pidx).len());
-            for i in 0..grm.prod(pidx).len() {
-                let argt = match grm.prod(pidx)[i] {
-                    Symbol::Rule(ref_ridx) => {
-                        let action_type = grm.actiontype(ref_ridx)
-                           .as_ref()
-                           .expect("actiontype should have been checked during complete_and_validate for this YaccKind");
-                        str::parse::<TokenStream>(action_type)?
-                    }
-                    Symbol::Token(_) => {
-                        let lexemet =
-                            str::parse::<TokenStream>(type_name::<LexerTypesT::LexemeT>())?;
-                        quote!(::std::result::Result<#lexemet, #lexemet>)
-                    }
-                };
-                let arg = format_ident!("{}arg_{}", ACTION_PREFIX, i + 1);
-                args.push(quote!(mut #arg: #argt));
-            }
-
-            // If this rule's `actiont` is `()` then Clippy will warn that the return type `-> ()`
-            // is pointless (which is true). We therefore avoid outputting a return type if actiont
-            // is the unit type.
-            let returnt = {
-                let actiont = grm.actiontype(grm.prod_to_rule(pidx)).as_ref().unwrap();
-                if actiont == "()" {
-                    None
-                } else {
-                    let actiont = str::parse::<TokenStream>(actiont)?;
-                    Some(quote!( -> #actiont))
-                }
-            };
-            let action_fn = format_ident!("{}action_{}", ACTION_PREFIX, usize::from(pidx));
-            let lexer_var = format_ident!("{}lexer", ACTION_PREFIX);
-            let span_var = format_ident!("{}span", ACTION_PREFIX);
-            let ridx_var = format_ident!("{}ridx", ACTION_PREFIX);
-            let storaget = str::parse::<TokenStream>(type_name::<LexerTypesT::StorageT>())?;
-            let lexertypest = str::parse::<TokenStream>(type_name::<LexerTypesT>())?;
-            let bind_parse_param = if !parse_param_unit {
-                Some(quote! {let _ = #parse_paramname;})
-            } else {
-                None
-            };
-
-            // Iterate over all $-arguments and replace them with their respective
-            // element from the argument vector (e.g. $1 is replaced by args[0]).
-            let pre_action = grm.action(pidx).as_ref().expect("action code should have been checked during complete_and_validate for this YaccKind");
-            let mut last = 0;
-            let mut outs = String::new();
-            loop {
-                match pre_action[last..].find('$') {
-                    Some(off) => {
-                        if pre_action[last + off..].starts_with("$$") {
-                            outs.push_str(&pre_action[last..last + off + "$".len()]);
-                            last = last + off + "$$".len();
-                        } else if pre_action[last + off..].starts_with("$lexer") {
-                            outs.push_str(&pre_action[last..last + off]);
-                            write!(outs, "{prefix}lexer", prefix = ACTION_PREFIX).ok();
-                            last = last + off + "$lexer".len();
-                        } else if pre_action[last + off..].starts_with("$span") {
-                            outs.push_str(&pre_action[last..last + off]);
-                            write!(outs, "{prefix}span", prefix = ACTION_PREFIX).ok();
-                            last = last + off + "$span".len();
-                        } else if last + off + 1 < pre_action.len()
-                            && pre_action[last + off + 1..].starts_with(|c: char| c.is_numeric())
-                        {
-                            outs.push_str(&pre_action[last..last + off]);
-                            write!(outs, "{prefix}arg_", prefix = ACTION_PREFIX).ok();
-                            last = last + off + "$".len();
-                        } else {
-                            unreachable!("action variables checked during complete_and_validate");
-                        }
-                    }
-                    None => {
-                        outs.push_str(&pre_action[last..]);
-                        break;
-                    }
-                }
-            }
-
-            let action_body = str::parse::<TokenStream>(&outs)?;
-            action_fns.extend(quote! {
-                #[allow(clippy::too_many_arguments)]
-                fn #action_fn #generics (
-                    #ridx_var: ::cfgrammar::RIdx<#storaget>,
-                    #lexer_var: &'lexer dyn ::lrpar::NonStreamingLexer<'input, #lexertypest>,
-                    #span_var: ::cfgrammar::Span,
-                    #parse_paramdef,
-                    #(#args,)*
-                ) #returnt
-                #where_clause
-                {
-                    #bind_parse_param
-                    #action_body
-                }
-            })
-        }
-        Ok(quote! {
-            #programs
-            #action_fns
-        })
-    }
-
     /// Return the `RIdx` of the %start rule in the grammar (which will not be the same as
     /// grm.start_rule_idx because the latter has an additional rule insert by cfgrammar
     /// which then calls the user's %start rule).
@@ -1635,7 +1496,7 @@ pub(crate) fn indent(indent: &str, s: &str) -> String {
     format!("{indent}{}\n", s.trim_end_matches('\n')).replace('\n', &format!("\n{}", indent))
 }
 
-fn make_generics(parse_generics: Option<&str>) -> Result<Generics, Box<dyn Error>> {
+pub(crate) fn make_generics(parse_generics: Option<&str>) -> Result<Generics, Box<dyn Error>> {
     if let Some(parse_generics) = parse_generics {
         let tokens = str::parse::<TokenStream>(parse_generics)?;
         match syn::parse2(quote!(<'lexer, 'input: 'lexer, #tokens>)) {
